+++ example/server-drizzle/package.json
@@ -9,8 +9,10 @@
   },
   "type": "module",
   "scripts": {
-    "db:generate": "NODE_OPTIONS='--env-file=.env' drizzle-kit generate --config drizzle.config.ts",
-    "db:push": "NODE_OPTIONS='--env-file=.env' drizzle-kit push --config drizzle.config.ts",
+    "db:check": "node --env-file=.env node_modules/drizzle-kit/bin.cjs check --config drizzle.config.ts",
+    "db:generate": "node --env-file=.env node_modules/drizzle-kit/bin.cjs generate --config drizzle.config.ts",
+    "db:migrate": "node --env-file=.env node_modules/drizzle-kit/bin.cjs migrate --config drizzle.config.ts",
+    "db:push": "node --env-file=.env node_modules/drizzle-kit/bin.cjs push --config drizzle.config.ts",
     "db:seed": "node --no-warnings --experimental-specifier-resolution=node --loader ts-node/esm --env-file .env src/drizzle/seed.tsx",
     "dev": "NODE_ENV=development node_modules/.bin/nodemon -q -I --exec node --no-warnings --experimental-specifier-resolution=node --loader ts-node/esm --env-file .env src/index.tsx",
     "dev:setup": "vp run db:push && vp run db:seed",
